--- src/main/java/frc/robot/subsystems/Intake.java
@@ -1,18 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
-public class Intake extends SubsystemBase {
-  /** Creates a new Intake. */
-  public Intake() {}
-
-  @Override
-  public void periodic() {
-    // This method will be called once per scheduler run
-    //adding this so commit can work (this can be deleted)
-  }
-}
