+++ src/main/java/frc/robot/util/NavX.java
@@ -5,18 +5,17 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.util.sensors;
+package frc.robot.util;
 
 import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj.SPI;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 /**
  * Add your docs here.
  */
 public class NavX {
     private final AHRS m_ahrs;
-    private double yawOffset = 0.0;
+    private double yawOffset;
 
     public NavX() {
         m_ahrs = new AHRS(SPI.Port.kMXP);
@@ -50,12 +49,5 @@ public class NavX {
     public float getRoll() {
         //Standard method
         return m_ahrs.getRoll();
-    }
-
-    public void periodic() {
-        SmartDashboard.putNumber("NavX Yaw", getYaw());
-        SmartDashboard.putNumber("NavX Yaw Rate", getYawRate());
-        SmartDashboard.putNumber("NavX Pitch", getPitch());
-        SmartDashboard.putNumber("NavX Roll", getRoll());
     }
 }
